+++ transaction/waitForConfirmation.go
@@ -1,4 +1,4 @@
-package future
+package transaction
 
 import (
 	"context"